Allow for render objects to be removed/added
Both renderers can now handle any objects that were
created at the start of the application to be removed and
then re-added. Support does *not* exist for adding any
number of objects, this might work in OpenGL, but
definitely not in Vulkan due to the initial pool size.

Pressing '0' will remove/re-add the grid object to test
that this feature works as intended.