Clean up uniform initialization
There are now three booleans for each shader uniform.
A "need" boolean is true when that shader has a uniform of that type.
  - This will cause the renderer to find the location of that uniform
An "enabled" boolean is true when that uniform is toggled on.
  - This will cause the renderer to send the data to the GPU
A "generate" boolean is true when that uniform needs something to
be generated, like a texture or cubemap.