Add first pass at reflection probes
In this commit reflection probes will capture their
surroundings, but only of objects that use forward
rendering, and everything is flipped vertically.

There is now a need for post-processing in order to
apply gamma correction and tone-mapping after all
passes complete. This can no longer happen inside
each shader as objects are rendered to a cubemap,
which is then rendered again.