Generate depth buffer for reflection probe cubemaps
Reflection probes are working! Well, only in GL and they
only render forward rendered objects. But it's a start.

A lot of time was spent figuring out FBOs and RBOs and
how to use them.

Some parts of this commit are still very unclean and hardcoded,
it will be cleaned up soon.