Improve deferred object reflection probe capture
Still not working!! (WHY?)
deferred_combine_cubemap isn't sampling input cubemap
samplers correctly for some unknown reason. The sample vectors
being used are the same as being used in the skybox shader,
the depth buffer read/write is set up correctly, and it's not being culled.

Not sure where to go with this one.