Add build system
After this commit you should be able to,when in the
scripts folder run "genie vs2015" and the necessary
project files will be generated. Vulkan headers and lib
have been added to make cloning the project more
streamlined.

Download GENie here if you don't already have it:
https://github.com/bkaradzic/GENie

Instructions soon to be added to the project readme