Ingore/add vertex attributes based on shader requirements
Vulkan shaders now specify which attributes they require and
mesh data respects this when loaded in. Data present in a mesh
but not required for a shader is ignored, and data not present in
a mesh but required in a shader is filled with default values.

Perhaps in future multiple copies of loaded in meshes will be
stored so that we don't load in the same mesh multiple times
when multiple shaders are being used.