Add prefab support
Objects can now be serialized to an individual
file to prevent duplicating the same fields for each
instance of an object. Once spawned the obejcts
bare no relation to one another.

Adding new objects is now easier since the
ToString and FromString functions use an array of
strings which only needs to be set in one place.

FindFIlesInDirectory is now friendlier. Different
parameter formats are handled properly (dir string
ending with a slash, file type beginning with dot)