Second pass on rotation gizmo functionality
Rotations are now handled as if there was a string between the
mouse-plane intersection point and the gizmo center. As you move
the mouse around the center the number of wraps is stored in
`m_RotationGizmoWrapCount`. This is all handled in
`FlexEngine::CalculateDeltaRotationFromGizmoDrag`.

Fade out axes when being faced head-on. There is still some issues
with this since we aren't sorting the three meshes based on
distance so the x-axis will always write to the depth buffer first,
causing a blank spot when facing it directly.