Edit render pass definitions & remove world pos from GBuffer
New render pass order:
1. Deferred pass to offscreen GBuffer & (new) offscreen depth
2. Deferred combine pass to swapchain color & depth
3. Blit offscreen depth to swapchain depth
4. Forward pass
5. UI

World position is no longer a component of the GBuffer, it will be determined from depth (which is now being passed into deferred combine) Currently using a stub in the shader.

SSAO framebuffers, passes, & shaders have been added in this commit but are non-functional and currently being skipped.

Main work in this commit was getting resources to be in the right layout at the right time. VkAttachmentDescription::finalLayout is quite handy for that!